+++ pkg/tempo/config.go
@@ -5,14 +5,18 @@ import (
 	"errors"
 	"fmt"
 	"io/ioutil"
+	"time"
 
+	"github.com/grafana/agent/pkg/tempo/noopreceiver"
 	"github.com/grafana/agent/pkg/tempo/promsdprocessor"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/spanmetricsprocessor"
 	prom_config "github.com/prometheus/common/config"
 	"github.com/spf13/viper"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configmodels"
 	"go.opentelemetry.io/collector/exporter/otlpexporter"
+	"go.opentelemetry.io/collector/exporter/prometheusexporter"
 	"go.opentelemetry.io/collector/processor/attributesprocessor"
 	"go.opentelemetry.io/collector/processor/batchprocessor"
 	"go.opentelemetry.io/collector/receiver/jaegerreceiver"
@@ -20,6 +24,11 @@ import (
 	"go.opentelemetry.io/collector/receiver/opencensusreceiver"
 	"go.opentelemetry.io/collector/receiver/otlpreceiver"
 	"go.opentelemetry.io/collector/receiver/zipkinreceiver"
+)
+
+const (
+	spanMetricsPipelineName    = "metrics/spanmetrics"
+	defaultSpanMetricsExporter = "prometheus"
 )
 
 // Config controls the configuration of Tempo trace pipelines.
@@ -73,6 +82,9 @@ type InstanceConfig struct {
 
 	// prom service discovery
 	ScrapeConfigs []interface{} `yaml:"scrape_configs,omitempty"`
+
+	// SpanMetricsProcessor: https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/main/processor/spanmetricsprocessor/README.md
+	SpanMetrics *SpanMetricsConfig `yaml:"spanmetrics,omitempty"`
 }
 
 const (
@@ -143,7 +155,16 @@ func (c *RemoteWriteConfig) UnmarshalYAML(unmarshal func(interface{}) error) err
 	return nil
 }
 
-// exporter builds an OTel exporter from PushConfig
+// SpanMetricsConfig controls the configuration of spanmetricsprocessor and the related metrics exporter.
+type SpanMetricsConfig struct {
+	LatencyHistogramBuckets []time.Duration                  `yaml:"latency_histogram_buckets,omitempty"`
+	Dimensions              []spanmetricsprocessor.Dimension `yaml:"dimensions,omitempty"`
+
+	// Configuration for Prometheus exporter: https://github.com/open-telemetry/opentelemetry-collector/blob/7d7ae2eb34/exporter/prometheusexporter/README.md.
+	MetricsExporter map[string]interface{} `yaml:"metrics_exporter,omitempty"`
+}
+
+// exporter builds an OTel exporter from RemoteWriteConfig
 func exporter(remoteWriteConfig RemoteWriteConfig) (map[string]interface{}, error) {
 	if len(remoteWriteConfig.Endpoint) == 0 {
 		return nil, errors.New("must have a configured a backend endpoint")
@@ -251,8 +272,6 @@ func (c *InstanceConfig) otelConfig() (*configmodels.Config, error) {
 		exportersNames = append(exportersNames, name)
 	}
 
-	otelMapStructure["exporters"] = exporters
-
 	// processors
 	processors := map[string]interface{}{}
 	processorNames := []string{}
@@ -276,24 +295,50 @@ func (c *InstanceConfig) otelConfig() (*configmodels.Config, error) {
 		processorNames = append(processorNames, "batch")
 	}
 
-	otelMapStructure["processors"] = processors
+	if c.SpanMetrics != nil {
+		// Configure the metrics exporter.
+		exporters[defaultSpanMetricsExporter] = c.SpanMetrics.MetricsExporter
+
+		processorNames = append(processorNames, "spanmetrics")
+		processors["spanmetrics"] = map[string]interface{}{
+			"metrics_exporter":          defaultSpanMetricsExporter,
+			"latency_histogram_buckets": c.SpanMetrics.LatencyHistogramBuckets,
+			"dimensions":                c.SpanMetrics.Dimensions,
+		}
+	}
 
 	// receivers
-	otelMapStructure["receivers"] = c.Receivers
 	receiverNames := []string{}
 	for name := range c.Receivers {
 		receiverNames = append(receiverNames, name)
 	}
 
+	pipelines := map[string]interface{}{
+		"traces": map[string]interface{}{
+			"exporters":  exportersNames,
+			"processors": processorNames,
+			"receivers":  receiverNames,
+		},
+	}
+
+	if c.SpanMetrics != nil {
+		// Insert a noop receiver in the metrics pipeline.
+		// Added to pass validation requiring at least one receiver in a pipeline.
+		c.Receivers[noopreceiver.TypeStr] = nil
+
+		pipelines[spanMetricsPipelineName] = map[string]interface{}{
+			"receivers": []string{noopreceiver.TypeStr},
+			"exporters": []string{defaultSpanMetricsExporter},
+		}
+	}
+
+	otelMapStructure["exporters"] = exporters
+	otelMapStructure["processors"] = processors
+	otelMapStructure["receivers"] = c.Receivers
+
 	// pipelines
 	otelMapStructure["service"] = map[string]interface{}{
-		"pipelines": map[string]interface{}{
-			"traces": map[string]interface{}{
-				"exporters":  exportersNames,
-				"processors": processorNames,
-				"receivers":  receiverNames,
-			},
-		},
+		"pipelines": pipelines,
 	}
 
 	// now build the otel configmodel from the mapstructure
@@ -330,6 +375,7 @@ func tracingFactories() (component.Factories, error) {
 		otlpreceiver.NewFactory(),
 		opencensusreceiver.NewFactory(),
 		kafkareceiver.NewFactory(),
+		noopreceiver.NewFactory(),
 	)
 	if err != nil {
 		return component.Factories{}, err
@@ -337,6 +383,7 @@ func tracingFactories() (component.Factories, error) {
 
 	exporters, err := component.MakeExporterFactoryMap(
 		otlpexporter.NewFactory(),
+		prometheusexporter.NewFactory(),
 	)
 	if err != nil {
 		return component.Factories{}, err
@@ -346,6 +393,7 @@ func tracingFactories() (component.Factories, error) {
 		batchprocessor.NewFactory(),
 		attributesprocessor.NewFactory(),
 		promsdprocessor.NewFactory(),
+		spanmetricsprocessor.NewFactory(),
 	)
 	if err != nil {
 		return component.Factories{}, err